adding app column in response and removing app_comments, app_responses data entry and updating app column in response for past records
