--- src/main/kotlin/uk/gov/justice/digital/hmpps/managingprisonerappsapi/dto/response/AppResponsePrisoner.kt
@@ -1,30 +1,0 @@
-package uk.gov.justice.digital.hmpps.managingprisonerappsapi.dto.response
-
-import com.fasterxml.jackson.annotation.JsonFormat
-import uk.gov.justice.digital.hmpps.managingprisonerappsapi.model.AppStatus
-import java.time.LocalDateTime
-import java.util.*
-
-data class AppResponsePrisoner<X, Y>(
-  val id: UUID?,
-  val reference: String?,
-  val assignedGroup: X,
-  val applicationType: ApplicationTypeResponse,
-  val genericForm: Boolean,
-  val applicationGroup: ApplicationGroupResponse,
-  @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'")
-  val requestedDate: LocalDateTime,
-  @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'")
-  val createdDate: LocalDateTime,
-  val createdBy: String,
-  @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'")
-  val lastModifiedDate: LocalDateTime?,
-  val lastModifiedBy: String?,
-  val requests: List<Map<String, Any>>,
-  val requestedBy: Y,
-  val requestedByFirstName: String,
-  val requestedByLastName: String,
-  val status: AppStatus,
-  val establishmentId: String,
-  val reason: String?,
-)
