+++ src/main/kotlin/uk/gov/justice/digital/hmpps/managingprisonerappsapi/model/Response.kt
@@ -20,7 +20,9 @@ data class Response(
   val decision: Decision,
   val createdDate: LocalDateTime,
   val createdBy: String,
+  val app: UUID,
 ) {
+
   override fun equals(other: Any?): Boolean {
     if (this === other) return true
     if (javaClass != other?.javaClass) return false
