+++ src/main/resources/db/migration/dev/V039__add_app_column_in_response_and_update_data.sql
@@ -1,0 +1,4 @@
+ALTER TABLE if exists response
+    ADD column if not exists app uuid;
+UPDATE response
+SET app = (SELECT app_responses.app_id FROM app_responses WHERE response.id = app_responses.responses);
